refactor(text response): move files
